@@ -18,17 +18,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.1.2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  matrix: ^0.5.5
+  # Matrix
+  matrix: ^0.7.0-nullsafety.5
+  olm: ^2.0.0
+
+  # Database
   hive: ^2.0.4
   hive_flutter: ^1.1.0
-  flutter_olm: ^1.1.1
+  
   #flutter_openssl_crypto: ^0.0.1
 
 
@@ -78,15 +82,17 @@ dependencies:
   #firebase_messaging: ^7.0.3
   #flutter_local_notifications: ^3.0.2
   # Flushbar
-  flushbar: ^1.10.4
+  another_flushbar: ^1.10.25
   #dialogs
   adaptive_dialog: ^0.10.0+5
   
 
   package_info_plus: ^1.0.0
   
-  minestrix_chat: 
+  minestrix_chat:
+    #path: ./minestrix-chat/
     git: https://gitlab.com/minestrix/minestrix-chat.git
+  collection: ^1.15.0-nullsafety.4
       
 dev_dependencies:
   flutter_test:
