@@ -6,11 +6,11 @@ import 'package:minestrix/utils/minestrix/minestrixRoom.dart';
 import 'package:minestrix_chat/partials/matrix_user_image.dart';
 
 class PostWriterModal extends StatelessWidget {
-  PostWriterModal({Key key, @required this.sroom}) : super(key: key);
-  final MinestrixRoom sroom;
+  PostWriterModal({Key? key, required this.sroom}) : super(key: key);
+  final MinestrixRoom? sroom;
   @override
   Widget build(BuildContext context) {
-    MinestrixClient sclient = Matrix.of(context).sclient;
+    MinestrixClient sclient = Matrix.of(context).sclient!;
     return Card(
         child: Padding(
       padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 15),
@@ -18,7 +18,7 @@ class PostWriterModal extends StatelessWidget {
         children: [
           MatrixUserImage(
               client: sclient,
-              url: sclient.userRoom.user.avatarUrl,
+              url: sclient.userRoom!.user!.avatarUrl,
               width: 48,
               thumnail: true,
               height: 48),
@@ -37,7 +37,7 @@ class PostWriterModal extends StatelessWidget {
                       children: [
                         Icon(Icons.edit),
                         SizedBox(width: 10),
-                        Expanded(child: Text("Write a post as " + sroom.name)),
+                        Expanded(child: Text("Write a post as " + sroom!.name!)),
                       ],
                     ),
                   )))
