@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 
-class LoginTitle extends StatelessWidget {
-  const LoginTitle({
-    Key key,
-  }) : super(key: key);
+class MinestrixTitle extends StatelessWidget {
+  const MinestrixTitle({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
