@@ -1,5 +1,5 @@
 import 'dart:async';
-import 'package:flushbar/flushbar.dart';
+import 'package:another_flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
@@ -10,9 +10,9 @@ import 'package:minestrix/utils/minestrix/minestrixClient.dart';
 import 'package:minestrix/utils/platforms_info.dart';
 
 class Matrix extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
 
-  Matrix({Key key, this.child}) : super(key: key);
+  Matrix({Key? key, this.child}) : super(key: key);
 
   @override
   MatrixState createState() => MatrixState();
@@ -20,7 +20,7 @@ class Matrix extends StatefulWidget {
   /// Returns the (nearest) Client instance of your application.
   static MatrixState of(BuildContext context) {
     var newState =
-        (context.dependOnInheritedWidgetOfExactType<_InheritedMatrix>()).data;
+        context.dependOnInheritedWidgetOfExactType<_InheritedMatrix>()!.data!;
     newState.context = context;
     return newState;
   }
@@ -29,16 +29,16 @@ class Matrix extends StatefulWidget {
 class MatrixState extends State<Matrix> {
   final log = Logger("MatrixState");
 
-  MinestrixClient sclient;
+  MinestrixClient? sclient;
   @override
-  BuildContext context;
+  late BuildContext context;
 
-  StreamSubscription<KeyVerification> onKeyVerificationRequestSub;
-  StreamSubscription<EventUpdate> onEvent;
+  StreamSubscription<KeyVerification>? onKeyVerificationRequestSub;
+  StreamSubscription<EventUpdate>? onEvent;
 
   @deprecated
   Future<void> connect() async {
-    sclient.onLoginStateChanged.stream.listen((LoginState loginState) {
+    sclient!.onLoginStateChanged.stream.listen((LoginState loginState) {
       print("LoginState: ${loginState.toString()}");
     });
   }
@@ -50,7 +50,7 @@ class MatrixState extends State<Matrix> {
   }
 
   void initMatrix() {
-    final Set verificationMethods = <KeyVerificationMethod>{
+    final Set<KeyVerificationMethod> verificationMethods = <KeyVerificationMethod>{
       KeyVerificationMethod.numbers
     };
 
@@ -65,12 +65,12 @@ class MatrixState extends State<Matrix> {
     sclient = MinestrixClient(clientName,
         enableE2eeRecovery: true, verificationMethods: verificationMethods);
 
-    print("logged: " + sclient.isLogged().toString());
+    print("logged: " + sclient!.isLogged().toString());
     print("[ widget ] : store");
     _initWithStore();
     print("[ widget ] : register");
 
-    onKeyVerificationRequestSub ??= sclient.onKeyVerificationRequest.stream
+    onKeyVerificationRequestSub ??= sclient!.onKeyVerificationRequest.stream
         .listen((KeyVerification request) async {
       print("KeyVerification");
       print(request.deviceId);
@@ -103,15 +103,15 @@ class MatrixState extends State<Matrix> {
       }*/
     });
 
-    onEvent ??= sclient.onEvent.stream
+    onEvent ??= sclient!.onEvent.stream
         .where((event) =>
             [EventTypes.Message, EventTypes.Encrypted]
                 .contains(event.content['type']) &&
-            event.content['sender'] != sclient.userID)
+            event.content['sender'] != sclient!.userID)
         .listen((EventUpdate eventUpdate) async {
       // we should react differently depending on wether the event is a smatrix one or not...
       // get event object
-      Room room = sclient.getRoomById(eventUpdate.roomID);
+      Room room = sclient!.getRoomById(eventUpdate.roomID)!;
       Event event = Event.fromJson(eventUpdate.content, room);
 
       // don't throw a notification for old events
@@ -120,11 +120,11 @@ class MatrixState extends State<Matrix> {
           0) {
         // check if it is a minestrix event or a message
         // This method works only for already recognised SRooms
-        bool isSRoom = sclient.srooms.containsKey(eventUpdate.roomID);
+        bool isSRoom = sclient!.srooms.containsKey(eventUpdate.roomID);
         if (isSRoom) {
-          Profile profile = await sclient.getUserFromRoom(room);
+          Profile profile = await sclient!.getUserFromRoom(room);
           Flushbar(
-            title: "New post from " + profile.displayName,
+            title: "New post from " + profile.displayName!,
             message: event.body,
             duration: Duration(seconds: 3),
             flushbarPosition: FlushbarPosition.TOP,
@@ -137,9 +137,9 @@ class MatrixState extends State<Matrix> {
           if (margin_left < 8) margin_left = 8;
           Flushbar(
             margin: EdgeInsets.only(bottom: 8, right: 8, left: margin_left),
-            borderRadius: 8,
+            borderRadius: BorderRadius.all(Radius.circular(8)),
             maxWidth: mWidth,
-            title: event.sender.displayName + "@" + room.name,
+            title: event.sender.displayName! + "@" + room.name,
             dismissDirection: FlushbarDismissDirection.HORIZONTAL,
             icon: Icon(Icons.info, color: Colors.white),
             message: event.body,
@@ -155,13 +155,13 @@ class MatrixState extends State<Matrix> {
   }
 
   void _initWithStore() async {
-    var initLoginState = sclient.onLoginStateChanged.stream.first;
+    var initLoginState = sclient!.onLoginStateChanged.stream.first;
     try {
-      sclient.init();
+      sclient!.init();
 
       final firstLoginState = await initLoginState;
       if (firstLoginState == LoginState.loggedIn) {
-        await sclient.initSMatrix();
+        await sclient!.initSMatrix();
       } else {
         log.warning("Not logged in");
       }
@@ -180,23 +180,23 @@ class MatrixState extends State<Matrix> {
   @override
   Widget build(BuildContext context) {
     log.info("build");
-    return _InheritedMatrix(data: this, child: widget.child);
+    return _InheritedMatrix(data: this, child: widget.child!);
   }
 }
 
 class _InheritedMatrix extends InheritedWidget {
-  final MatrixState data;
+  final MatrixState? data;
 
-  _InheritedMatrix({Key key, this.data, Widget child})
+  _InheritedMatrix({Key? key, this.data, required Widget child})
       : super(key: key, child: child);
 
   @override
   bool updateShouldNotify(_InheritedMatrix old) {
-    var update = old.data.sclient.accessToken != data.sclient.accessToken ||
-        old.data.sclient.userID != data.sclient.userID ||
-        old.data.sclient.deviceID != data.sclient.deviceID ||
-        old.data.sclient.deviceName != data.sclient.deviceName ||
-        old.data.sclient.homeserver != data.sclient.homeserver;
+    var update = old.data!.sclient!.accessToken != data!.sclient!.accessToken ||
+        old.data!.sclient!.userID != data!.sclient!.userID ||
+        old.data!.sclient!.deviceID != data!.sclient!.deviceID ||
+        old.data!.sclient!.deviceName != data!.sclient!.deviceName ||
+        old.data!.sclient!.homeserver != data!.sclient!.homeserver;
     return update;
   }
 }
