@@ -11,8 +11,8 @@ import 'package:minestrix_chat/partials/matrix_user_image.dart';
 class FriendsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final MinestrixClient sclient = Matrix.of(context).sclient;
-    List<User> users = sclient.userRoom.room
+    final MinestrixClient sclient = Matrix.of(context).sclient!;
+    List<User> users = sclient.userRoom!.room!
         .getParticipants()
         .where((User u) => u.membership == Membership.join)
         .toList();
@@ -32,30 +32,30 @@ class FriendsPage extends StatelessWidget {
             suggestionsCallback: (pattern) async {
               var ur = await sclient.searchUserDirectory(pattern);
 
-              List<User> following = List<User>.empty();
-              await sclient.following.forEach((key, MinestrixRoom sroom) {
+              List<User?> following = List<User?>.empty();
+              sclient.following.forEach((key, MinestrixRoom sroom) {
                 following.add(sroom.user);
               });
 
               return ur.results
                   .where((element) =>
                       following.firstWhere(
-                          (friend) => friend.id == element.userId,
+                          (friend) => friend!.id == element.userId,
                           orElse: () => null) ==
                       null)
                   .toList(); // exclude current friends
             },
-            itemBuilder: (context, suggestion) {
+            itemBuilder: (context, dynamic suggestion) {
               Profile profile = suggestion;
               return ListTile(
                 leading: profile.avatarUrl == null
                     ? Icon(Icons.person)
                     : MatrixUserImage(client: sclient, url: profile.avatarUrl),
-                title: Text(profile.displayName),
+                title: Text(profile.displayName!),
                 subtitle: Text(profile.userId),
               );
             },
-            onSuggestionSelected: (suggestion) async {
+            onSuggestionSelected: (dynamic suggestion) async {
               Profile p = suggestion;
               await sclient.addFriend(p.userId);
             },
@@ -78,9 +78,9 @@ class FriendsPage extends StatelessWidget {
                             Row(
                               children: [
                                 MatrixUserImage(
-                                    client: sclient, url: sm.user.avatarUrl),
+                                    client: sclient, url: sm.user!.avatarUrl),
                                 SizedBox(width: 10),
-                                Text(sm.user.displayName),
+                                Text(sm.user!.displayName!),
                               ],
                             ),
                             Row(
@@ -89,12 +89,12 @@ class FriendsPage extends StatelessWidget {
                                     icon:
                                         Icon(Icons.check, color: Colors.green),
                                     onPressed: () async {
-                                      await sm.room.join();
+                                      await sm.room!.join();
                                     }),
                                 IconButton(
                                     icon: Icon(Icons.delete, color: Colors.red),
                                     onPressed: () async {
-                                      await sm.room.leave();
+                                      await sm.room!.leave();
                                     }),
                               ],
                             ),
